Added a GraphQL and JSON setup
Introduced graphql and graphql-mode. The graphql package provide a DSL
for generating graphql queries in elisp, and the request module allow
us to execute the queries against a GraphQL server. The graphql-mode
package provide syntax highlighting in .graphql-files, and support for
executing commands as well, based on the graphql specification.

I added json-mode and json-reformat to the mix as well. I generally
don't like JSON but at least this package makes it easier to work
with; especially the reformat command that I've bound to "C-c TAB"; I
need to make this command check if a region is selected, as it require
that to function.