Made json-mode work again
It relied on the package json-snatcher for some
reason. `json-snatcher` seems to work once in a while. I suspect it
has a couple of rough edges. I also played around with json-reformat,
and it seems to have edge cases where it will out-right delete data
when it performs it reformat.

I might remove this entire blob; the thought of reformatting json is
nice though, especially when working with restclient mode. I don't
care enough about json to build my own mode.