Add yaml-mode
This is needed for when I need to poke around in docker-compose.yml
files when I work with Docker.

I also added a todo about the docker set environment function I did; I
did this as a todo in the code because I cannot create issues on my
github repo--because it is a clone, I should probably just upgrade it
to be my own since I've changed almost anything since I cloned it back
in the day.